# new/protocol/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Invity",
    "description": "Volvo API.",
    "version": "0.1"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/project": {
      "get": {
        "summary": "remove last response",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean",
                  "description": "is user a mentor of the project"
                },
                "leader": {
                  "type": "boolean",
                  "description": "is user a leader of the project"
                },
                "member": {
                  "type": "boolean",
                  "description": "is user a member of the project"
                },
                "favourite": {
                  "type": "boolean",
                  "description": "is user's favourite project"
                },
                "recruitment": {
                  "type": "boolean",
                  "description": "is recruitment open"
                },
                "invited": {
                  "type": "boolean",
                  "description": "has been user invited to the project"
                },
                "applied": {
                  "type": "boolean",
                  "description": "has user applied to the project"
                },
                "project": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "depiction": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string",
                      "format": "path"
                    }
                  }
                },
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "skill": {
                        "type": "string",
                        "description": "title of skill connected with role"
                      },
                      "members": {
                        "type": "array",
                        "description": "all project members involved in role",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "avatar": {
                              "type": "string",
                              "format": "path"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      },
                      "link": {
                        "type": "string",
                        "format": "url"
                      }
                    }
                  }
                },
                "closure": {
                  "type": "object",
                  "properties": {
                    "closed": {
                      "type": "boolean",
                      "description": "is project closed"
                    },
                    "state": {
                      "type": "boolean",
                      "description": "was closure request sent by the leader"
                    },
                    "explanation": {
                      "type": "string",
                      "description": "the reason why the leader want to close the project"
                    }
                  }
                },
                "response": {
                  "type": "object",
                  "description": "message thet will be shown when user opens a project page; often reffered as 'send message'",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project"
        ]
      }
    },
...
    "/project/report": {
      "post": {
        "summary": "report project",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/report?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "explanation": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
...
    "/reports": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean"
                },
                "leader": {
                  "type": "boolean"
                },
                "closed": {
                  "type": "boolean"
                },
                "reports": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "description": "project reports",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "answers": {
                        "type": "array",
                        "description": "answers to 4 questions in report form",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "responses": {
                  "description": "responses from mentor",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "text": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project",
          "project/reports"
        ]
      }
    },
    "/reports/link": {
      "post": {
        "summary": "send link to mentor's email address; note user name (security purpose)",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /reports/link?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "link": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/reports"
        ]
      }
    },
    "/reports/send": {
      "post": {
        "summary": "send report to mentor",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /reports/send?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "string"
                },
                "answers": {
                  "type": "array",
                  "description": "answers to 4 questions in report form",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/reports"
        ]
      }
    },
    "/reports/update": {
      "post": {
        "summary": "update project reports",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /reports/update?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "reports": {
                  "type": "array"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/reports"
        ]
      }
    },
...
    "/team/report": {
      "post": {
        "summary": "report member",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /team/report",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "project": {
                  "type": "string"
                },
                "explanation": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/team"
        ]
      }
    },
...
  },
...
}